adjusted dragon spawning and returning to intro cutscene
added a pack away manager to interrupt all active entites before return to the intr oscene. Dragon prefab is bugged and wont register as grounded.

+++ Assets/GameFiles/refactored scrips/Conditions/HealthCondition.cs
@@ -17,7 +17,7 @@ public class HealthCondition : BaseCondition
     public override void Initialize(Entity entity)
     {
         ownerEntity = entity;
-        Debug.Log("current health percent: " + ownerEntity.healthSystem.currentHealth / ownerEntity.healthSystem.maxHealth);
+        Debug.Log("current health percent: " + (float)ownerEntity.healthSystem.currentHealth / (float)ownerEntity.healthSystem.maxHealth);
     }
     public override void ConditionUpdate()
     {
